requirements.txt
fastapi
uvicorn
python-multipart
playwright
openai
python-dotenv
numpy
